code distribution and cosmetic changes: usart_driver.c usart_handler.c
